use inferno for signature help, and workaround the way render works w/inferno
