+++ src/bootstrap/index.html
@@ -13,6 +13,11 @@
         src: url('../assets/roboto-mono.ttf');
       }
 
+      #plugins > div {
+        width: 100%;
+        height: 100%;
+      }
+
       html {
         --font: 'Roboto Mono Veonim';
         --font-size: 14;
